refactor(broker): extract LogStoragePartitionStep
